implemented Token rpository interface
